backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
Flask>=2.0.1
flask-cors>=3.0.10
tensorflow>=2.15.0
Pillow>=9.0.0
numpy>=1.19.5
